[fp-rs] Renaming folder windows_ble to windows, moving common to folder
