+++ fastpair/rust/src/bluetooth/windows/adapter.rs
