+++ fastpair/rust/src/bluetooth/windows/device.rs
